fixed axios service error handling

+++ src/axios.service.ts
@@ -1,4 +1,4 @@
-import axios, { AxiosResponse } from "axios";
+import axios, { AxiosError, AxiosResponse } from "axios";
 
 // We need to abstract this and the baseUrl of server in "axios.service.ts" to an env file
 export const baseServerUrl = "https://cocktail-party-server.herokuapp.com";
@@ -17,7 +17,7 @@ export function getAuthToken() {
 
 authToken = localStorage.getItem('cocktailPartyAccessToken') || "";
 
-export function get(endpoint: string, params: any, resHandler?: (res: AxiosResponse) => void, errHandler?: (err: any) => void) {
+export function get(endpoint: string, params: any, resHandler?: (res: AxiosResponse) => void, errHandler?: (err: AxiosError) => void) {
     axios.get(
         `${baseServerUrl}${endpoint}`,
         {
@@ -26,19 +26,19 @@ export function get(endpoint: string, params: any, resHandler?: (res: AxiosRespo
                 "authorization": "Bearer " + authToken
             }
         }
-    ).then(resHandler).catch((error: { status: number; }) => {
-        if (error.status === 401) {
+    ).then(resHandler).catch((error: AxiosError) => {
+        if (error.response && error.response.status === 401) {
             localStorage.clear();
         }
         
-        (errHandler || ((err: any) => {
+        (errHandler !== undefined ? errHandler : ((err: AxiosError) => {
             // TO DO - write some error dialog / snackbar to display this information
             console.log(err);
         }))(error);
     });
 }
 
-export function post(endpoint: string, data: any, params?: any, resHandler?: (res: AxiosResponse) => void, errHandler?: (err: any) => void) {
+export function post(endpoint: string, data: any, params?: any, resHandler?: (res: AxiosResponse) => void, errHandler?: (err: AxiosError) => void) {
     axios.post(
         `${baseServerUrl}${endpoint}`,
         data,
@@ -48,19 +48,19 @@ export function post(endpoint: string, data: any, params?: any, resHandler?: (re
                 "authorization": "Bearer " + authToken
             }
         }
-    ).then(resHandler).catch((error: { status: number; }) => {
-        if (error.status === 401) {
+    ).then(resHandler).catch((error: AxiosError) => {
+        if (error.response && error.response.status === 401) {
             localStorage.clear();
         }
         
-        (errHandler || ((err: any) => {
+        (errHandler !== undefined ? errHandler : ((err: AxiosError) => {
             // TO DO - write some error dialog / snackbar to display this information
             console.log(err);
         }))(error);
     });
 }
 
-export function del(endpoint: string, params?: any, resHandler?: (res: AxiosResponse) => void, errHandler?: (err: any) => void) {
+export function del(endpoint: string, params?: any, resHandler?: (res: AxiosResponse) => void, errHandler?: (err: AxiosError) => void) {
     axios.delete(
         `${baseServerUrl}${endpoint}`,
         {
@@ -69,14 +69,13 @@ export function del(endpoint: string, params?: any, resHandler?: (res: AxiosResp
                 "authorization": "Bearer " + authToken
             }
         }
-    ).then(resHandler).catch((error: { status: number; }) => {
-        if (error.status === 401) {
+    ).then(resHandler).catch((error: AxiosError) => {
+        if (error.response && error.response.status === 401) {
             localStorage.clear();
         }
 
         
-        
-        (errHandler || ((err: any) => {
+        (errHandler !== undefined ? errHandler : ((err: AxiosError) => {
             // TO DO - write some error dialog / snackbar to display this information
             console.log(err);
         }))(error);
